chore(utils): move test-utils to separate file
