+++ __tests__/gists.test.ts
@@ -5,7 +5,7 @@
  */
 
 import * as stripAnsi from 'strip-ansi'
-import { runCmd } from './testUtils'
+import { runCmd } from '../src/test-utils'
 
 describe('E2E: Gist Module Test', () => {
     it('List Gists `gh gi`', done => {
