+++ __tests__/milestone.test.ts
@@ -4,7 +4,7 @@
  * SPDX-License-Identifier: BSD-3-Clause
  */
 
-import { runCmd } from './testUtils'
+import { runCmd } from '../src/test-utils'
 
 describe('E2E: Milestone Module Test', () => {
     it('List Milestones `gh ms --list`', done => {
